Changing Api\RequestHandler to load customer orders directly
Prior to this change, the RequestHandler would load customer orders by
setting the customer id in the session and then calling a method on a
frontend block class. Adding the id to the session would log in the
customer, even if they were checking out as a guest. This commit removes the
session and frontend block classes and uses the order collection factory
to load orders for the customer directly. The customer order data can
then be used for the api call and the customer session is not modified.