Validating last 4 cc digits before adding to API request
In some cases, we do not have the last four digits of the credit card
stored in our payment table and a placeholder like 'xxxx' is used
instead. The NoFraud API returns an error if the last4 are not all
numbers. This commit adds validation that checks if the last4 pulled
from the database is actually four digits before adding it to the API
request.